Moving PeirForce to correct namespace
