+++ ETABS_oM/Results/PierForce.cs
@@ -29,7 +29,7 @@ using System.Threading.Tasks;
 using BH.oM.Base;
 using BH.oM.Structure.Elements;
 
-namespace BH.oM.Adapters.ETABS.Elements
+namespace BH.oM.Adapters.ETABS.Results
 {
     public class PierForce : BH.oM.Structure.Results.BarForce
     {
